% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqm.R
\encoding{UTF-8}
\name{freqm}
\alias{freqm}
\title{Tris à plats multiples pondérable}
\usage{
freqm(
  table,
  racine_var,
  calc = "all",
  poids = NULL,
  total = T,
  exclude = NULL,
  transpose = F,
  verbose = T
)
}
\arguments{
\item{table}{table de données}

\item{racine_var}{racine des variables issues de quest. à choix multiple}

\item{calc}{choix du type de calcul : fréquence et/ou pourcentages = "n", "\%", "all"}

\item{poids}{variable contenant les poids}

\item{total}{Si F, n'ajoute pas le total}

\item{exclude}{vecteur : Valeur(s) à exclure ; ex : c("-NC-",NA)}

\item{transpose}{Si T, présentation en transposée}

\item{verbose}{Si F, ne copie pas le tableau dans la console,
quand le résultat est envoyé dans un objet par exemple}
}
\value{
Un tableau de fréquences et/ou pourcentage
}
\description{
Tris à plats multiples pondérable
}
