% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqp.R
\encoding{UTF-8}
\name{freqp}
\alias{freqp}
\title{Effectifs bruts, pondérés selon 1 ou + poids dans un même tableau}
\usage{
freqp(
  table,
  var,
  total = T,
  exclude = "",
  transpose = F,
  poids = NULL,
  verbose = T,
  brut = T,
  freq = T,
  pourc = T
)
}
\arguments{
\item{table}{table de données}

\item{var}{variable}

\item{total}{Si F, n'ajoute pas le total}

\item{exclude}{vecteur : valeur(s) à exclure ; ex : c("-NC-",NA)}

\item{transpose}{Si T, présentation en transposée}

\item{poids}{variable(s) contenant les poids}

\item{verbose}{Si F, ne copie pas le tableau dans la console,
quand le résultat est envoyé dans un objet par exemple}

\item{brut}{Si F, ne présente pas les valeurs sans pondération}

\item{freq}{Si F, ne présente pas les valeurs de fréquence}

\item{pourc}{Si F, ne présente pas les valeurs de pourcentages}
}
\value{
Un tableau de fréquences et/ou pourcentage
}
\description{
Effectifs bruts, pondérés selon 1 ou + poids dans un même tableau
}
